code improvements and corrections

+++ src/main/java/com/example/quiz_world/quiz/quizCategory/QuizCategoryService.java
@@ -2,6 +2,7 @@ package com.example.quiz_world.quiz.quizCategory;
 
 import com.example.quiz_world.exception.ExistsException;
 import com.example.quiz_world.mapper.MapperEntity;
+import com.example.quiz_world.quiz.quiz.QuizRepository;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -12,19 +13,29 @@ import java.util.Optional;
 @Service
 public class QuizCategoryService {
     private final QuizCategoryRepository quizCategoryRepository;
+    private final QuizRepository quizRepository;
     private final MapperEntity mapperEntity;
 
-    public QuizCategoryService(QuizCategoryRepository quizCategoryRepository, MapperEntity mapperEntity) {
+    public QuizCategoryService(QuizCategoryRepository quizCategoryRepository, QuizRepository quizRepository,
+                               MapperEntity mapperEntity) {
         this.quizCategoryRepository = quizCategoryRepository;
+        this.quizRepository = quizRepository;
         this.mapperEntity = mapperEntity;
     }
 
-    public QuizCategory createQuizCategory(QuizCategory quizCategory) {
-        Optional<QuizCategory> existingCategory = quizCategoryRepository.findByNameIgnoreCase(quizCategory.getName());
+    public QuizCategoryDTO createQuizCategory(QuizCategoryRequest quizCategoryRequest) {
+        Optional<QuizCategory> existingCategory = quizCategoryRepository.findByNameIgnoreCase
+                (quizCategoryRequest.getCategoryName());
         if (existingCategory.isPresent()) {
             throw new ExistsException("Category exists");
         }
-        return quizCategoryRepository.save(quizCategory);
+
+        QuizCategory quizCategory = new QuizCategory();
+        quizCategory.setName(quizCategoryRequest.getCategoryName());
+
+        quizCategoryRepository.save(quizCategory);
+
+        return mapperEntity.mapQuizCategoryToQuizCategoryDTO(quizCategory);
     }
 
     public List<QuizCategoryDTO> findAllQuizCategories() {
@@ -37,25 +48,35 @@ public class QuizCategoryService {
 
     @Transactional
     public void deleteQuizCategoryById(Long categoryId) {
-        QuizCategory quizCategory = quizCategoryRepository.findById(categoryId).orElseThrow(() -> new EntityNotFoundException("Quiz category not found"));
+        QuizCategory quizCategory = quizCategoryRepository.findById(categoryId).orElseThrow(() ->
+                new EntityNotFoundException("Quiz category not found"));
+
+        if (quizRepository.existsByQuizCategoryId(categoryId)) {
+            throw new IllegalStateException("Cannot delete category as it is associated with existing quizzes.");
+        }
 
         quizCategoryRepository.delete(quizCategory);
     }
 
     @Transactional
     public void deleteAllQuizCategories() {
-        List<QuizCategory> quizCategoryList = quizCategoryRepository.findAll();
-        if (quizCategoryList.isEmpty()) {
-            throw new EntityNotFoundException("Not found categories in Quiz");
+        if (quizRepository.findFirstBy().isPresent()) {
+            throw new IllegalArgumentException("Cannot delete categories because there are quizzes associated with them.");
         }
         quizCategoryRepository.deleteAll();
     }
 
-    public QuizCategory updateQuizCategory(Long idCategory, QuizCategory updateCategory) {
-        QuizCategory presentCategory = quizCategoryRepository.findById(idCategory).orElseThrow(() -> new EntityNotFoundException("Category not found"));
+    public QuizCategoryDTO updateQuizCategory(Long idCategory, QuizCategoryRequest quizCategoryRequest) {
+        QuizCategory presentCategory = quizCategoryRepository.findById(idCategory).orElseThrow(() ->
+                new EntityNotFoundException("Category not found"));
+        if (quizCategoryRepository.existsByName(quizCategoryRequest.getCategoryName())) {
+            throw new IllegalArgumentException("Category name already exists");
+        }
+        presentCategory.setName(quizCategoryRequest.getCategoryName());
 
-        presentCategory.setName(updateCategory.getName());
-        return quizCategoryRepository.save(presentCategory);
+        quizCategoryRepository.save(presentCategory);
+
+        return mapperEntity.mapQuizCategoryToQuizCategoryDTO(presentCategory);
     }
 }
